타이포에러 import문 최적화

+++ src/main/webapp/WEB-INF/views/list-customers.jsp
@@ -1,5 +1,5 @@
 <%@ page language="java" contentType="text/html; charset=ISO-8859-1"
-         pageEncoding="ISO-8859-1" isELIgnored="false" %>
+         pageEncoding="ISO-8859-1"%>
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
 <!DOCTYPE html>
 <html>
@@ -37,32 +37,54 @@
                     </tr>
 
                     <!-- loop over and print our customers -->
+                    <%--          <c:forEach var="tempCustomer" items="${customers}">--%>
+
+                    <%--            <!-- construct an "update" link with customer id -->--%>
+                    <%--            <c:url var="updateLink" value="/updateForm">--%>
+                    <%--              <c:param name="customerId" value="${tempCustomer.id}"/>--%>
+                    <%--            </c:url>--%>
+
+                    <%--            <!-- construct an "delete" link with customer id -->--%>
+                    <%--            <c:url var="deleteLink" value="${delete}">--%>
+                    <%--              <c:param name="customerId" value="${tempCustomer.id}" />--%>
+                    <%--            </c:url>--%>
+
+                    <%--            <tr>--%>
+                    <%--              <td>${tempCustomer.firstName}</td>--%>
+                    <%--              <td>${tempCustomer.lastName}</td>--%>
+                    <%--              <td>${tempCustomer.email}</td>--%>
+
+                    <%--              <td>--%>
+                    <%--                <!-- display the update link --> <a href="${updateLink}">Update</a>--%>
+                    <%--                | <a href="${deleteLink}"--%>
+                    <%--                     onclick="if (!(confirm('Are you sure you want to delete this customer?'))) return false">Delete</a>--%>
+                    <%--              </td>--%>
+
+                    <%--            </tr>--%>
+
+                    <%--          </c:forEach>--%>
+
                     <c:forEach var="tempCustomer" items="${customers}">
-
-                        <!-- construct an "update" link with customer id -->
-                        <c:url var="updateLink" value="/customer/updateForm">
-                            <c:param name="customerId" value="${tempCustomer.id}" />
-                        </c:url>
-
-                        <!-- construct an "delete" link with customer id -->
-                        <c:url var="deleteLink" value="/customer/delete">
-                            <c:param name="customerId" value="${tempCustomer.id}" />
-                        </c:url>
-
                         <tr>
                             <td>${tempCustomer.firstName}</td>
                             <td>${tempCustomer.lastName}</td>
                             <td>${tempCustomer.email}</td>
-
                             <td>
-                                <!-- display the update link --> <a href="${updateLink}">Update</a>
-                                | <a href="${deleteLink}"
-                                     onclick="if (!(confirm('Are you sure you want to delete this customer?'))) return false">Delete</a>
+                                <!-- "Update" ??? ?? form -->
+                                <form action="/customer/updateForm" method="get">
+                                    <input type="hidden" name="customerId" value="${tempCustomer.id}" />
+                                    <input type="submit" value="Update" />
+                                </form>
+                                <form action="/customer/delete" method="get">
+                                    <input type="hidden" name="customerId" value="${tempCustomer.id}" />
+                                    <input type="submit" value="Delete" />
+                                </form>
+                                <!-- "Delete" ?? (?? ?? ??) -->
+                                    <%--                <a href="${deleteLink}" onclick="if (!(confirm('Are you sure you want to delete this customer?'))) return false">Delete</a>--%>
                             </td>
-
                         </tr>
-
                     </c:forEach>
+
 
                 </table>
